Deprecate property tests
Never had any significant difference between default and property-based tests. Both have the same behavior, but, property-based showed a `~` instead of `-` to tell us this is a property of the object.

Honestly, this is useless, and since I'm upgrading to a new major version (3), I'm removing this.

Our test description's should be clear enough to give us the right context to know if this is a property, context-bound, or whatever other test cases.